faceRecog/requirements.txt
numpy>=1.24.0,<2.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
insightface==0.7.3
opencv-python-headless==4.8.1.78
onnxruntime==1.16.3
